requirements.txt
streamlit==1.19.0
openai==0.27.0
easyocr==1.5.0
opencv-python==4.5.5.64
pillow==9.0.0
numpy==1.23.3
matplotlib==3.7.0
pydantic==1.10.2
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
